Switch to storing track as a LineString
