+++ src/helpers/calculations/bearingCalc.ts
@@ -1,5 +1,5 @@
 import { Calculation, GraphDataset, GraphDatum } from "../../components/GraphModal";
-import { Feature, MultiPoint, Position } from 'geojson';
+import { Feature, LineString, Position } from 'geojson';
 import { isTemporal } from "../trackCalculations";
 import * as turf from "@turf/turf";
 
@@ -8,7 +8,7 @@ import * as turf from "@turf/turf";
  * then return the coordinates of the point at that index
  */
 const nearestPoint = (feature: Feature, time: number): Position | undefined => {
-  const geom = feature.geometry as MultiPoint
+  const geom = feature.geometry as LineString
   const times = feature.properties?.times as number[]
   const index = times.findIndex((t) => t >= time)
   if (index === -1) {
@@ -29,7 +29,7 @@ export const bearingCalc: Calculation = {
       return []
     }
     const nonBaseTemporal = temporal.filter((feature) => feature.id !== baseId)
-    const baseGeom = baseTrack.geometry as MultiPoint
+    const baseGeom = baseTrack.geometry as LineString
     return nonBaseTemporal.map((feature) => {
       const name = feature.properties?.name || feature.id
