+++ src/helpers/tests/featureIsVisibleAtTime.test.ts
@@ -27,7 +27,7 @@ const track: Feature = {
   },
   geometry: {
     coordinates: [[1.0, 1.0], [2.0, 2.0], [3.0, 3.0], [4.0, 4.0], [5.0, 5.0], [6.0, 6.0]],
-    type: 'MultiPoint'
+    type: 'LineString'
   },
   id: 'f-1'
 }
